+++ src/manifest.json
@@ -34,7 +34,7 @@
     {
       "extension_ids": ["hfojakphgokgpbnejoobfamojbgolcbo"],
       "matches": ["<all_urls>"],
-      "resources": ["ogkush.js", "main.js", "libs/*", "assets/*", "util/*", "ctxpage/*"]
+      "resources": ["ogkush.js", "main.js", "libs/*", "assets/*", "util/*", "ctxpage/*", "ctxcontent/*"]
     }
   ]
 }
