+++ src/manifestv2.json
@@ -15,7 +15,7 @@
     }
   ],
 
-  "web_accessible_resources": ["ogkush.js", "libs/*", "assets/*", "util/*", "ctxpage/*"],
+  "web_accessible_resources": ["ogkush.js", "libs/*", "assets/*", "util/*", "ctxpage/*", "ctxcontent/*"],
 
   "browser_action": {
     "default_title": "Ogame Infinity",
